Assignment/ass2-backend/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==0.24
MarkupSafe==1.0
Werkzeug==0.14.1
flask-restplus==0.12.1
Flask-Cors==3.0.6
Pillow==5.3.0
